Re-do renaming of repo_relocation.
Symbols weren't being lifted, so this shouldn't get a leading
underscore.
